拆分类，转为多文件结构
